% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arxiv_storage.R
\name{arxiv_load_csv}
\alias{arxiv_load_csv}
\title{Load arXiv records from CSV}
\usage{
arxiv_load_csv(path = NULL)
}
\arguments{
\item{path}{File path for CSV. If NULL, uses default package data dir.}
}
\value{
Tibble (possibly empty).
}
\description{
Load arXiv records from CSV
}
