% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arxiv_storage.R
\name{arxiv_load_meta}
\alias{arxiv_load_meta}
\title{Load update metadata (watermark)}
\usage{
arxiv_load_meta(path = NULL)
}
\arguments{
\item{path}{File path for JSON/DPUT. If NULL, uses default package data dir.}
}
\value{
Named list (possibly empty).
}
\description{
Load update metadata (watermark)
}
\keyword{internal}
