% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arxiv_storage.R
\name{arxiv_save_meta}
\alias{arxiv_save_meta}
\title{Save update metadata (watermark)}
\usage{
arxiv_save_meta(meta, path = NULL)
}
\arguments{
\item{meta}{Named list. Stored as JSON.}

\item{path}{File path for JSON. If NULL, uses default package data dir.}
}
\value{
Path to written JSON.
}
\description{
Save update metadata (watermark)
}
\keyword{internal}
