% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_shiny.R
\name{run_app}
\alias{run_app}
\title{Run arxivThreatIntel Shiny GUI}
\usage{
run_app(host = "127.0.0.1", port = 8080, launch_browser = TRUE)
}
\arguments{
\item{host}{Shiny host (default 127.0.0.1)}

\item{port}{Shiny port (default 8080)}

\item{launch_browser}{Open browser automatically}
}
\description{
Simple GUI to update arXiv cache, tag papers with TI topics, and explore results.
}
