% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_topics.R
\name{ti_load_topics}
\alias{ti_load_topics}
\title{Load topic dictionary (patterns) from package extdata}
\usage{
ti_load_topics(path = NULL)
}
\arguments{
\item{path}{Optional path to CSV with columns: topic, pattern.}
}
\value{
Tibble(topic, pattern)
}
\description{
Load topic dictionary (patterns) from package extdata
}
