% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_primary_topic.R
\name{ti_primary_topic_stats}
\alias{ti_primary_topic_stats}
\title{Get primary topic statistics}
\usage{
ti_primary_topic_stats(x)
}
\arguments{
\item{x}{Tibble with 'primary_topic' column}
}
\value{
Tibble with columns: primary_topic, n, pct
}
\description{
Get primary topic statistics
}
