% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_stats.R
\name{ti_topic_freq}
\alias{ti_topic_freq}
\title{Compute topic frequency table}
\usage{
ti_topic_freq(x)
}
\arguments{
\item{x}{Tagged tibble/data.frame (output of ti_tag()) with list-column 'topics'.}
}
\value{
Tibble with columns: topic, n
}
\description{
Compute topic frequency table
}
