% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_stats.R
\name{ti_topic_trend_weekly}
\alias{ti_topic_trend_weekly}
\title{Compute topic trend by week}
\usage{
ti_topic_trend_weekly(x)
}
\arguments{
\item{x}{Tagged tibble/data.frame with columns 'published' and list-column 'topics'.}
}
\value{
Tibble with columns: week, topic, n
}
\description{
Compute topic trend by week
}
